优化条件判断

+++ src/main/java/com/ruoyi/project/system/dict/service/DictTypeServiceImpl.java
@@ -133,7 +133,7 @@ public class DictTypeServiceImpl implements IDictTypeService
     {
         Long dictId = StringUtils.isNull(dict.getDictId()) ? -1L : dict.getDictId();
         DictType dictType = dictTypeMapper.checkDictTypeUnique(dict.getDictType());
-        if (StringUtils.isNotNull(dictType) && dictType.getDictId() != dictId)
+        if (StringUtils.isNotNull(dictType) && dictType.getDictId().longValue() != dictId.longValue())
         {
             return UserConstants.DICT_TYPE_NOT_UNIQUE;
         }
